Implemented the concurrent renderer, it is faster than the single threaded version.
