+++ src/ConcurrentRenderer.cpp
@@ -4,8 +4,10 @@
 #include <format>
 
 #include "Vector.h"
+#include "Utilities.h"
 
 void RenderBlock(Scene& scene, Camera& camera, Image& target, int startX, int startY, int blockSize);
+Vector RayColor(const Ray& ray, const Scene& scene, int depth);
 
 ConcurrentRenderer::ConcurrentRenderer(Scene& scene, int threadCount, int blockSize) : renderThreads(threadCount), scene(scene), blockSize(blockSize)
 {
@@ -29,13 +31,46 @@ ConcurrentRenderer::~ConcurrentRenderer()
 	renderThreads.~ThreadPool();
 }
 
+constexpr auto samples = 16;
+constexpr auto maxDepth = 20;
+
 void RenderBlock(Scene& scene, Camera& camera, Image& target, int startX, int startY, int blockSize)
 {
+	int width = target.GetWidth();
+	int height = target.GetHeight();
+
 	for (int y = startY; y < startY + blockSize; y++) {
 		for (int x = startX; x < startX + blockSize; x++) {
+			Vector color;
+			for (int i = 0; i < samples; i++) {
+				double u = (double(x) + Random()) / (width - 1);
+				double v = (double(y) + Random()) / (height - 1);
+				color += RayColor(camera.GetRay(u, v), scene, maxDepth);
+			}
 
+			double sampleFactor = 1.0 / samples;
+			target.WritePixel(x, y, color * sampleFactor);
 		}
 	}
+}
 
-	std::this_thread::sleep_for(std::chrono::duration<double>(0.001));
+Vector RayColor(const Ray& ray, const Scene& scene, int depth) {
+	if (depth <= 0) {
+		return Vector();
+	}
+
+	HitRecord hit;
+	if (scene.IsHitByRay(ray, 0.0000001, infinity, hit)) {
+
+		Ray scatter;
+		Vector attenuation;
+
+		if (hit.material->ScatterRay(ray, hit, attenuation, scatter)) {
+			return attenuation * RayColor(scatter, scene, depth - 1);
+		}
+
+		return Vector();
+	}
+	double t = 0.5 * ((ray.direction).GetNormal().y + 1);
+	return Vector(1.0, 1.0, 1.0) * (1.0 - t) + Vector(0.0, 0.0, 1.0) * t;
 }
