+++ src/Main.cpp
@@ -23,27 +23,6 @@ constexpr auto samples = 25;
 constexpr auto maxDepth = 20;
 constexpr double aspectRatio = (double)WIDTH / HEIGHT;
 
-Vector RayColor(const Ray& ray, const Scene& scene, int depth) {
-	if (depth <= 0) {
-		return Vector();
-	}
-
-	HitRecord hit;
-	if (scene.IsHitByRay(ray, 0.0000001, infinity, hit)) {
-
-		Ray scatter;
-		Vector attenuation;
-		
-		if (hit.material->ScatterRay(ray, hit, attenuation, scatter)) {
-			return attenuation * RayColor(scatter, scene, depth - 1);
-		}
-
-		return Vector();
-	}
-	double t = 0.5 * ((ray.direction).GetNormal().y + 1);
-	return Vector(1.0, 1.0, 1.0) * (1.0 - t) + Vector(0.0, 0.0, 1.0) * t;
-}
-
 int main() {
 	Image image(WIDTH, HEIGHT);
 	Vector lookFrom = { 7, 7, 2 };
@@ -75,13 +54,13 @@ int main() {
 	scene.AddObject(backSphere);
 
 	auto t1 = std::chrono::high_resolution_clock::now();
-	ConcurrentRenderer renderer(scene, 10, 40);
+	ConcurrentRenderer renderer(scene, 7, 40);
 	renderer.RenderCurrentScene(camera, image);
 	auto t2 = std::chrono::high_resolution_clock::now();
 	std::chrono::duration<double> diff = t2 - t1;
 	std::cout << std::format("[INFO] Render took {:.5f} seconds\n", diff.count());
 
-	t1 = std::chrono::high_resolution_clock::now();
+	/*t1 = std::chrono::high_resolution_clock::now();
 	for (int y = 0; y < HEIGHT; y++) {
 		for (int x = 0; x < WIDTH; x++) {
 			Vector color;
@@ -100,6 +79,7 @@ int main() {
 	t2 = std::chrono::high_resolution_clock::now();
 	diff = t2 - t1;
 	std::cout << std::format("[INFO] Render took {:.5f} seconds\n", diff.count());
+	*/
 
 	std::filesystem::current_path(std::filesystem::path("images"));
 	image.WriteImageTofile("simpleSphere.ppm");
