+++ src/objects/Hittable.h
@@ -11,7 +11,7 @@ struct HitRecord
 	Vector point;
 	Vector normal;
 	double t = 0.0;
-	bool isFrontFace;
+	bool isFrontFace = false;
 	std::shared_ptr<class Material> material;
 
 	inline void SetNormal(const Ray& ray, const Vector& outwardNormal)
